+++ src/main/java/com/atlauncher/gui/tabs/settings/GeneralSettingsTab.java
@@ -36,6 +36,7 @@ import com.atlauncher.App;
 import com.atlauncher.builders.HTMLBuilder;
 import com.atlauncher.constants.Constants;
 import com.atlauncher.constants.UIConstants;
+import com.atlauncher.data.AddModRestriction;
 import com.atlauncher.data.Language;
 import com.atlauncher.gui.components.JLabelWithHover;
 import com.atlauncher.utils.ComboItem;
@@ -51,6 +52,7 @@ public class GeneralSettingsTab extends AbstractSettingsTab {
     private final JComboBox<ComboItem<String>> dateFormat;
     private final JComboBox<ComboItem<Integer>> selectedTabOnStartup;
     private final JComboBox<ComboItem<String>> defaultModPlatform;
+    private final JComboBox<ComboItem<AddModRestriction>> addModRestriction;
     private final JCheckBox sortPacksAlphabetically;
     private final JCheckBox showPackNameAndVersion;
     private final JCheckBox keepLauncherOpen;
@@ -58,7 +60,6 @@ public class GeneralSettingsTab extends AbstractSettingsTab {
     private final JCheckBox enableTrayIcon;
     private final JCheckBox enableDiscordIntegration;
     private JCheckBox enableFeralGamemode;
-    private final JCheckBox disableAddModRestrictions;
     private final JCheckBox disableCustomFonts;
     private final JCheckBox rememberWindowSizePosition;
     private final JCheckBox useNativeFilePicker;
@@ -231,6 +232,40 @@ public class GeneralSettingsTab extends AbstractSettingsTab {
 
         add(defaultModPlatform, gbc);
 
+        // Add Mod Restrictions
+
+        gbc.gridx = 0;
+        gbc.gridy++;
+        gbc.insets = UIConstants.LABEL_INSETS;
+        gbc.anchor = GridBagConstraints.BASELINE_TRAILING;
+
+        JLabelWithHover addModRestrictionsLabel = new JLabelWithHover(GetText.tr("Add Mod Restrictions") + ":",
+                HELP_ICON, GetText.tr("What restrictions should be in place when adding mods from a mod platform."));
+
+        add(addModRestrictionsLabel, gbc);
+
+        gbc.gridx++;
+        gbc.insets = UIConstants.FIELD_INSETS;
+        gbc.anchor = GridBagConstraints.BASELINE_LEADING;
+        addModRestriction = new JComboBox<>();
+        addModRestriction.addItem(
+                new ComboItem<>(AddModRestriction.STRICT, GetText.tr("Only show mods for current Minecraft version")));
+        addModRestriction.addItem(new ComboItem<>(AddModRestriction.LAX,
+                GetText.tr("Show mods for the current major Minecraft version (eg: 1.16.x)")));
+        addModRestriction
+                .addItem(new ComboItem<>(AddModRestriction.NONE, GetText.tr("Show mods for all Minecraft versions")));
+
+        for (int i = 0; i < addModRestriction.getItemCount(); i++) {
+            ComboItem<AddModRestriction> item = addModRestriction.getItemAt(i);
+
+            if (item.getValue() == App.settings.addModRestriction) {
+                addModRestriction.setSelectedIndex(i);
+                break;
+            }
+        }
+
+        add(addModRestriction, gbc);
+
         // Sort Packs Alphabetically
 
         gbc.gridx = 0;
@@ -381,26 +416,6 @@ public class GeneralSettingsTab extends AbstractSettingsTab {
             add(enableFeralGamemode, gbc);
         }
 
-        // Disable CurseForge Minecraft version restrictions
-
-        gbc.gridx = 0;
-        gbc.gridy++;
-        gbc.insets = UIConstants.LABEL_INSETS;
-        gbc.anchor = GridBagConstraints.BASELINE_TRAILING;
-        JLabelWithHover disableAddModRestrictionsLabel = new JLabelWithHover(
-                GetText.tr("Disable Add Mod Restrictions?"), HELP_ICON,
-                new HTMLBuilder().center().split(100).text(GetText.tr(
-                        "This will allow you to disable the restrictions in place to prevent you from installing mods that are not for your current Minecraft version or loader. By disabling these restrictions, you can install any mod, so be sure that it's compatible with the Minecraft version and loader (if any) that you're on."))
-                        .build());
-        add(disableAddModRestrictionsLabel, gbc);
-
-        gbc.gridx++;
-        gbc.insets = UIConstants.CHECKBOX_FIELD_INSETS;
-        gbc.anchor = GridBagConstraints.BASELINE_LEADING;
-        disableAddModRestrictions = new JCheckBox();
-        disableAddModRestrictions.setSelected(App.settings.disableAddModRestrictions);
-        add(disableAddModRestrictions, gbc);
-
         // Disable custom fonts
 
         gbc.gridx = 0;
@@ -493,6 +508,7 @@ public class GeneralSettingsTab extends AbstractSettingsTab {
         App.settings.dateFormat = ((ComboItem<String>) dateFormat.getSelectedItem()).getValue();
         App.settings.selectedTabOnStartup = ((ComboItem<Integer>) selectedTabOnStartup.getSelectedItem()).getValue();
         App.settings.defaultModPlatform = ((ComboItem<String>) defaultModPlatform.getSelectedItem()).getValue();
+        App.settings.addModRestriction = ((ComboItem<AddModRestriction>) addModRestriction.getSelectedItem()).getValue();
         App.settings.sortPacksAlphabetically = sortPacksAlphabetically.isSelected();
         App.settings.showPackNameAndVersion = showPackNameAndVersion.isSelected();
         App.settings.keepLauncherOpen = keepLauncherOpen.isSelected();
@@ -506,7 +522,6 @@ public class GeneralSettingsTab extends AbstractSettingsTab {
             App.settings.enableFeralGamemode = false;
         }
 
-        App.settings.disableAddModRestrictions = disableAddModRestrictions.isSelected();
         App.settings.disableCustomFonts = disableCustomFonts.isSelected();
         App.settings.rememberWindowSizePosition = rememberWindowSizePosition.isSelected();